feat: refactor registry management to use ConfigsRegistry and improve structure
