+++ src/registries/mod.rs
@@ -1,0 +1,2 @@
+pub mod configs_registry;
+pub mod package_registry;
