+++ src/tasks/configs_registry.rs
@@ -1,19 +1,18 @@
-use crate::cli::{RegistryActions, RegistryArgs};
+use crate::cli::{ConfigsRegistryActions, ConfigsRegistryArgs};
 use crate::logger::log;
-use crate::registry::{Category, LinkRegistry, RegistryEntry, TargetPath};
+use crate::registries::configs_registry::{ConfigsRegistry, RegistryEntry, TargetPath};
 use crate::utils::paths::get_registry_path;
-use std::str::FromStr;
 
 /// Run the registry management command
-pub fn run(args: RegistryArgs) {
+pub fn run(args: ConfigsRegistryArgs) {
     match args.action {
-        RegistryActions::List {
+        ConfigsRegistryActions::List {
             category,
             enabled_only,
         } => {
             list_entries(category, enabled_only);
         }
-        RegistryActions::Add {
+        ConfigsRegistryActions::Add {
             id,
             name,
             source,
@@ -23,10 +22,10 @@ pub fn run(args: RegistryArgs) {
         } => {
             add_entry(id, name, source, target, category, description);
         }
-        RegistryActions::Remove { id } => {
+        ConfigsRegistryActions::Remove { id } => {
             remove_entry(id);
         }
-        RegistryActions::Toggle { id, enable } => {
+        ConfigsRegistryActions::Toggle { id, enable } => {
             toggle_entry(id, enable);
         }
     }
@@ -35,7 +34,7 @@ pub fn run(args: RegistryArgs) {
 /// List registry entries
 fn list_entries(filter_category: Option<String>, enabled_only: bool) {
     let registry_path = get_registry_path();
-    let registry = match LinkRegistry::load_or_create(&registry_path) {
+    let registry = match ConfigsRegistry::load_or_create(&registry_path) {
         Ok(registry) => registry,
         Err(e) => {
             println!("❌ Failed to load registry: {}", e);
@@ -110,7 +109,7 @@ fn add_entry(
     description: Option<String>,
 ) {
     let registry_path = get_registry_path();
-    let mut registry = match LinkRegistry::load_or_create(&registry_path) {
+    let mut registry = match ConfigsRegistry::load_or_create(&registry_path) {
         Ok(registry) => registry,
         Err(e) => {
             println!("❌ Failed to load registry: {}", e);
@@ -179,7 +178,7 @@ fn add_entry(
 /// Remove an entry from the registry
 fn remove_entry(id: String) {
     let registry_path = get_registry_path();
-    let mut registry = match LinkRegistry::load_or_create(&registry_path) {
+    let mut registry = match ConfigsRegistry::load_or_create(&registry_path) {
         Ok(registry) => registry,
         Err(e) => {
             println!("❌ Failed to load registry: {}", e);
@@ -208,7 +207,7 @@ fn remove_entry(id: String) {
 /// Toggle an entry's enabled status
 fn toggle_entry(id: String, enable: bool) {
     let registry_path = get_registry_path();
-    let mut registry = match LinkRegistry::load_or_create(&registry_path) {
+    let mut registry = match ConfigsRegistry::load_or_create(&registry_path) {
         Ok(registry) => registry,
         Err(e) => {
             println!("❌ Failed to load registry: {}", e);
