+++ src/tasks/restore.rs
@@ -1,5 +1,5 @@
 use crate::logger::log;
-use crate::registry::LinkRegistry;
+use crate::registries::configs_registry::ConfigsRegistry;
 use crate::utils::paths::{get_backup_path, get_base_dirs, get_registry_path};
 use std::fs;
 use std::path::PathBuf;
@@ -19,7 +19,7 @@ pub fn run() {
 
     // Load the registry
     let registry_path = get_registry_path();
-    let registry = match LinkRegistry::load_or_create(&registry_path) {
+    let registry = match ConfigsRegistry::load_or_create(&registry_path) {
         Ok(registry) => registry,
         Err(e) => {
             println!("❌ Failed to load registry: {}", e);
